migrate変更、order完成

+++ db/migrate/20191212051227_create_orders.rb
@@ -2,7 +2,7 @@ class CreateOrders < ActiveRecord::Migration[5.2]
   def change
     create_table :orders do |t|
       t.integer :public_id
-      t.integer :tax_rate
+      t.float :tax_rate
       t.integer :delivery_charge
       t.integer :total_price
       t.integer :delivery_status, index: true, default: 0
@@ -11,6 +11,8 @@ class CreateOrders < ActiveRecord::Migration[5.2]
       t.string :address_name
       t.string :zip
       t.datetime :deleted_at
+
+
 
       t.timestamps
     end
